feat: dialog de confirmacao e responsividade para tela principal

+++ src/app/components/pages/gerenciamento/gerenciamento.component.ts
@@ -4,6 +4,8 @@ import { IUsuario } from '../../../Interfaces/Usuario.interface';
 import { NavBarComponent } from "../../nav-bar/nav-bar.component";
 import { ListaDeUsuariosComponent } from "./lista-de-usuarios/lista-de-usuarios.component";
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-gerenciamento',
@@ -16,7 +18,10 @@ export class GerenciamentoComponent implements OnInit {
 
   snackBar = inject(MatSnackBar);
 
-  constructor(private readonly _usuarioService: UsuarioService) { }
+  constructor(
+    private readonly _usuarioService: UsuarioService,
+    private readonly dialog: MatDialog
+  ) { }
 
   usuarios: IUsuario[] = [];
 
@@ -37,24 +42,32 @@ export class GerenciamentoComponent implements OnInit {
   }
 
   removerUsuario(usuarioParaRemover: IUsuario) {
-    if (usuarioParaRemover && usuarioParaRemover.id) {
-      this._usuarioService.removerUsuario(usuarioParaRemover.id).subscribe({
-        next: () => {
-          this.snackBar.open('Usuário removido com sucesso', 'Fechar', {
-            duration: 5000,
-          });
-          this.getUsers();
-        },
-        error: () => {
-          this.snackBar.open('Erro ao remover usuário', 'Fechar', {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { user: usuarioParaRemover }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        if (usuarioParaRemover && usuarioParaRemover.id) {
+          this._usuarioService.removerUsuario(usuarioParaRemover.id).subscribe({
+            next: () => {
+              this.snackBar.open('Usuário removido com sucesso', 'Fechar', {
+                duration: 5000,
+              });
+              this.getUsers();
+            },
+            error: () => {
+              this.snackBar.open('Erro ao remover usuário', 'Fechar', {
+                duration: 5000,
+              });
+            }
+          })
+        } else {
+          this.snackBar.open('Não foi possível encontrar um usuário', 'Fechar', {
             duration: 5000,
           });
         }
-      })
-    } else {
-      this.snackBar.open('Não foi possível encontrar um usuário', 'Fechar', {
-        duration: 5000,
-      });
-    }
+      }
+    }) 
   }
 }
